+++ src/utils/toast.tsx
@@ -1,0 +1,3 @@
+import { TinyReactToastManager } from "@hiogawa/tiny-toast";
+
+export const toast = new TinyReactToastManager();
